Added REP (0x20) to line 40
When the game loads in an enemy with a "Can't Concentrate" initial status effect, the P register is set to make the A register 8-bit. This script requires that the A register be 16-bit instead so I need to add REP (0x20) to change the P register before we execute any code.